Added OpenAPI schema definitions for rules management
