+++ x-pack/plugins/security_solution/common/api/detection_engine/model/error_schema.mock.ts
