+++ x-pack/plugins/security_solution/common/api/detection_engine/model/error_schema.schema.yaml
@@ -1,0 +1,33 @@
+openapi: 3.0.0
+info:
+  title: Error Schema
+  version: 'not applicable'
+paths: {}
+components:
+  schemas:
+    ErrorSchema:
+      type: object
+      required:
+        - error
+      properties:
+        id:
+          type: string
+        rule_id:
+          $ref: './rule_schema/common_attributes.schema.yaml#/components/schemas/RuleSignatureId'
+        list_id:
+          type: string
+          minLength: 1
+        item_id:
+          type: string
+          minLength: 1
+        error:
+          type: object
+          required:
+            - status_code
+            - message
+          properties:
+            status_code:
+              type: integer
+              minimum: 400
+            message:
+              type: string
