+++ x-pack/plugins/security_solution/common/api/detection_engine/model/error_schema.test.ts
@@ -5,13 +5,12 @@
  * 2.0.
  */
 
+import { exactCheck, foldLeftRight, getPaths } from '@kbn/securitysolution-io-ts-utils';
 import { left } from 'fp-ts/lib/Either';
 import { pipe } from 'fp-ts/lib/pipeable';
-
 import type { ErrorSchema } from './error_schema';
 import { errorSchema } from './error_schema';
-import { exactCheck, foldLeftRight, getPaths } from '@kbn/securitysolution-io-ts-utils';
-import { getErrorSchemaMock } from './error_schema.mocks';
+import { getErrorSchemaMock } from './error_schema.mock';
 
 describe('error_schema', () => {
   test('it should validate an error with a UUID given for id', () => {
