--- x-pack/plugins/security_solution/common/api/detection_engine/model/rule_schema/common_attributes/related_integrations.ts
@@ -1,73 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import * as t from 'io-ts';
-import { NonEmptyString } from '@kbn/securitysolution-io-ts-types';
-
-/**
- * Related integration is a potential dependency of a rule. It's assumed that if the user installs
- * one of the related integrations of a rule, the rule might start to work properly because it will
- * have source events (generated by this integration) potentially matching the rule's query.
- *
- *   NOTE: Proper work is not guaranteed, because a related integration, if installed, can be
- *   configured differently or generate data that is not necessarily relevant for this rule.
- *
- * Related integration is a combination of a Fleet package and (optionally) one of the
- * package's "integrations" that this package contains. It is represented by 3 properties:
- *
- *   - `package`: name of the package (required, unique id)
- *   - `version`: version of the package (required, semver-compatible)
- *   - `integration`: name of the integration of this package (optional, id within the package)
- *
- * There are Fleet packages like `windows` that contain only one integration; in this case,
- * `integration` should be unspecified. There are also packages like `aws` and `azure` that contain
- * several integrations; in this case, `integration` should be specified.
- *
- * @example
- * const x: RelatedIntegration = {
- *   package: 'windows',
- *   version: '1.5.x',
- * };
- *
- * @example
- * const x: RelatedIntegration = {
- *   package: 'azure',
- *   version: '~1.1.6',
- *   integration: 'activitylogs',
- * };
- */
-export type RelatedIntegration = t.TypeOf<typeof RelatedIntegration>;
-export const RelatedIntegration = t.exact(
-  t.intersection([
-    t.type({
-      package: NonEmptyString,
-      version: NonEmptyString,
-    }),
-    t.partial({
-      integration: NonEmptyString,
-    }),
-  ])
-);
-
-/**
- * Array of related integrations.
- *
- * @example
- * const x: RelatedIntegrationArray = [
- *   {
- *     package: 'windows',
- *     version: '1.5.x',
- *   },
- *   {
- *     package: 'azure',
- *     version: '~1.1.6',
- *     integration: 'activitylogs',
- *   },
- * ];
- */
-export type RelatedIntegrationArray = t.TypeOf<typeof RelatedIntegrationArray>;
-export const RelatedIntegrationArray = t.array(RelatedIntegration);
