+++ x-pack/plugins/security_solution/common/api/detection_engine/prebuilt_rules/get_prebuilt_rules_and_timelines_status/get_prebuilt_rules_and_timelines_status_route.schema.yaml
@@ -1,0 +1,61 @@
+openapi: 3.0.0
+info:
+  title: Prebuilt Rules Status API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/prepackaged/_status:
+    get:
+      operationId: GetPrebuiltRulesStatus
+      x-codegen-enabled: false
+      summary: Get the status of Elastic prebuilt rules
+      tags:
+        - Prebuilt Rules API
+      responses:
+        200:
+          description: Indicates a successful call
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/GetPrebuiltRulesStatusResponse'
+
+components:
+  schemas:
+    GetPrebuiltRulesStatusResponse:
+      type: object
+      properties:
+        rules_custom_installed:
+          type: integer
+          description: The total number of custom rules
+          minimum: 0
+        rules_installed:
+          type: integer
+          description: The total number of installed prebuilt rules
+          minimum: 0
+        rules_not_installed:
+          type: integer
+          description: The total number of available prebuilt rules that are not installed
+          minimum: 0
+        rules_not_updated:
+          type: integer
+          description: The total number of outdated prebuilt rules
+          minimum: 0
+        timelines_installed:
+          type: integer
+          description: The total number of installed prebuilt timelines
+          minimum: 0
+        timelines_not_installed:
+          type: integer
+          description: The total number of available prebuilt timelines that are not installed
+          minimum: 0
+        timelines_not_updated:
+          type: integer
+          description: The total number of outdated prebuilt timelines
+          minimum: 0
+      required:
+        - rules_custom_installed
+        - rules_installed
+        - rules_not_installed
+        - rules_not_updated
+        - timelines_installed
+        - timelines_not_installed
+        - timelines_not_updated
