+++ x-pack/plugins/security_solution/common/api/detection_engine/prebuilt_rules/install_prebuilt_rules_and_timelines/install_prebuilt_rules_and_timelines_route.schema.yaml
@@ -1,0 +1,46 @@
+openapi: 3.0.0
+info:
+  title: Install Prebuilt Rules API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/prepackaged:
+    put:
+      operationId: InstallPrebuiltRules
+      x-codegen-enabled: false
+      summary: Installs all Elastic prebuilt rules and timelines
+      tags:
+        - Prebuilt Rules API
+      responses:
+        200:
+          description: Indicates a successful call
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/InstallPrebuiltRulesResponse'
+
+components:
+  schemas:
+    InstallPrebuiltRulesResponse:
+      type: object
+      properties:
+        rules_installed:
+          type: integer
+          description: The number of rules installed
+          minimum: 0
+        rules_updated:
+          type: integer
+          description: The number of rules updated
+          minimum: 0
+        timelines_installed:
+          type: integer
+          description: The number of timelines installed
+          minimum: 0
+        timelines_updated:
+          type: integer
+          description: The number of timelines updated
+          minimum: 0
+      required:
+        - rules_installed
+        - rules_updated
+        - timelines_installed
+        - timelines_updated
