+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_actions/bulk_actions_route.ts
@@ -53,13 +53,9 @@ export enum BulkActionEditType {
 export type ThrottleForBulkActions = t.TypeOf<typeof ThrottleForBulkActions>;
 export const ThrottleForBulkActions = t.union([
   t.literal('rule'),
-  TimeDuration({
-    allowedDurations: [
-      [1, 'h'],
-      [1, 'd'],
-      [7, 'd'],
-    ],
-  }),
+  t.literal('1h'),
+  t.literal('1d'),
+  t.literal('7d'),
 ]);
 
 type BulkActionEditPayloadTags = t.TypeOf<typeof BulkActionEditPayloadTags>;
