+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_create_rules/bulk_create_rules_route.schema.yaml
@@ -1,0 +1,29 @@
+openapi: 3.0.0
+info:
+  title: Bulk Create API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/_bulk_create:
+    post:
+      operationId: CreateRulesBulk
+      x-codegen-enabled: false
+      deprecated: true
+      description: Creates new detection rules in bulk.
+      tags:
+        - Bulk API
+      requestBody:
+        description: A JSON array of rules, where each rule contains the required fields.
+        required: true
+        content:
+          application/json:
+            schema:
+              type: array
+              items:
+                $ref: '../../../model/rule_schema/rule_schemas.schema.yaml#/components/schemas/RuleCreateProps'
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                $ref: '../response_schema.schema.yaml#/components/schemas/BulkCrudRulesResponse'
