+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_delete_rules/bulk_delete_rules_route.schema.yaml
@@ -1,0 +1,34 @@
+openapi: 3.0.0
+info:
+  title: Bulk Delete API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/_bulk_delete:
+    delete:
+      operationId: DeleteRulesBulk
+      x-codegen-enabled: false
+      deprecated: true
+      description: Deletes multiple rules.
+      tags:
+        - Bulk API
+      requestBody:
+        description: A JSON array of `id` or `rule_id` fields of the rules you want to delete.
+        required: true
+        content:
+          application/json:
+            schema:
+              type: array
+              items:
+                type: object
+                properties:
+                  id:
+                    $ref: '../../../model/rule_schema/common_attributes.schema.yaml#/components/schemas/RuleObjectId'
+                  rule_id:
+                    $ref: '../../../model/rule_schema/common_attributes.schema.yaml#/components/schemas/RuleSignatureId'
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                $ref: '../response_schema.schema.yaml#/components/schemas/BulkCrudRulesResponse'
