+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_patch_rules/bulk_patch_rules_route.schema.yaml
@@ -1,0 +1,29 @@
+openapi: 3.0.0
+info:
+  title: Bulk Patch API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/_bulk_update:
+    patch:
+      operationId: PatchRulesBulk
+      x-codegen-enabled: false
+      deprecated: true
+      description: Updates multiple rules using the `PATCH` method.
+      tags:
+        - Bulk API
+      requestBody:
+        description: A JSON array of rules, where each rule contains the required fields.
+        required: true
+        content:
+          application/json:
+            schema:
+              type: array
+              items:
+                $ref: '../../../model/rule_schema/rule_schemas.schema.yaml#/components/schemas/RulePatchProps'
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                $ref: '../response_schema.schema.yaml#/components/schemas/BulkCrudRulesResponse'
