+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_patch_rules/bulk_patch_rules_route.ts
@@ -6,10 +6,10 @@
  */
 
 import * as t from 'io-ts';
-import { PatchRuleRequestBody } from '../../crud/patch_rule/patch_rule_route';
+import { RulePatchProps } from '../../../model';
 
 /**
  * Request body parameters of the API route.
  */
 export type BulkPatchRulesRequestBody = t.TypeOf<typeof BulkPatchRulesRequestBody>;
-export const BulkPatchRulesRequestBody = t.array(PatchRuleRequestBody);
+export const BulkPatchRulesRequestBody = t.array(RulePatchProps);
