+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/bulk_crud/bulk_update_rules/bulk_patch_rules_route.schema.yaml
@@ -1,0 +1,29 @@
+openapi: 3.0.0
+info:
+  title: Bulk Update API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/_bulk_update:
+    put:
+      operationId: UpdateRulesBulk
+      x-codegen-enabled: false
+      deprecated: true
+      description: Updates multiple rules using the `PUT` method.
+      tags:
+        - Bulk API
+      requestBody:
+        description: A JSON array where each element includes the `id` or `rule_id` field of the rule you want to update and the fields you want to modify.
+        required: true
+        content:
+          application/json:
+            schema:
+              type: array
+              items:
+                $ref: '../../../model/rule_schema/rule_schemas.schema.yaml#/components/schemas/RuleUpdateProps'
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                $ref: '../response_schema.schema.yaml#/components/schemas/BulkCrudRulesResponse'
