+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/create_rule/create_rule_route.schema.yaml
@@ -1,0 +1,25 @@
+openapi: 3.0.0
+info:
+  title: Create Rule API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules:
+    post:
+      operationId: CreateRule
+      x-codegen-enabled: false
+      description: Create a single detection rule
+      tags:
+        - Rules API
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              $ref: '../../../model/rule_schema/rule_schemas.schema.yaml#/components/schemas/RuleCreateProps'
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                $ref: '../../../model/rule_schema/rule_schemas.schema.yaml#/components/schemas/RuleResponse'
