+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/delete_rule/delete_rule_route.schema.yaml
@@ -1,0 +1,32 @@
+openapi: 3.0.0
+info:
+  title: Delete Rule API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules:
+    delete:
+      operationId: DeleteRule
+      x-codegen-enabled: false
+      description: Deletes a single rule using the `rule_id` or `id` field.
+      tags:
+        - Rules API
+      parameters:
+        - name: id
+          in: query
+          required: false
+          description: The rule's `id` value.
+          schema:
+            $ref: '../../../model/rule_schema/common_attributes.schema.yaml#/components/schemas/RuleSignatureId'
+        - name: rule_id
+          in: query
+          required: false
+          description: The rule's `rule_id` value.
+          schema:
+            $ref: '../../../model/rule_schema/common_attributes.schema.yaml#/components/schemas/RuleObjectId'
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                $ref: '../../../model/rule_schema/rule_schemas.schema.yaml#/components/schemas/RuleResponse'
