+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/crud/patch_rule/patch_rule_route.ts
@@ -6,7 +6,7 @@
  */
 
 import type * as t from 'io-ts';
-import { RulePatchProps, RuleResponse, ThresholdRulePatchProps } from '../../../model';
+import { RulePatchProps, RuleResponse } from '../../../model';
 
 /**
  * Request body parameters of the API route.
@@ -14,9 +14,6 @@ import { RulePatchProps, RuleResponse, ThresholdRulePatchProps } from '../../../
  */
 export type PatchRuleRequestBody = RulePatchProps;
 export const PatchRuleRequestBody = RulePatchProps;
-
-export type ThresholdPatchRuleRequestBody = ThresholdRulePatchProps;
-export const ThresholdPatchRuleRequestBody = ThresholdRulePatchProps;
 
 export const PatchRuleResponse = RuleResponse;
 export type PatchRuleResponse = t.TypeOf<typeof PatchRuleResponse>;
