+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/find_rules/find_rules_route.ts
@@ -7,7 +7,7 @@
 
 import * as t from 'io-ts';
 import { DefaultPerPage, DefaultPage } from '@kbn/securitysolution-io-ts-alerting-types';
-import type { PerPage, Page, RuleResponse } from '../../model';
+import type { RuleResponse } from '../../model';
 import { SortOrder, queryFilter, fields } from '../../model';
 
 export type FindRulesSortField = t.TypeOf<typeof FindRulesSortField>;
@@ -45,11 +45,6 @@ export const FindRulesRequestQuery = t.exact(
     per_page: DefaultPerPage, // defaults to 20
   })
 );
-
-export type FindRulesRequestQueryDecoded = Omit<FindRulesRequestQuery, 'per_page'> & {
-  page: Page;
-  per_page: PerPage;
-};
 
 export interface FindRulesResponse {
   page: number;
