+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/import_rules/import_rules_route.schema.yaml
@@ -1,0 +1,54 @@
+openapi: 3.0.0
+info:
+  title: Import Rules API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/rules/_import:
+    summary: Imports rules from an `.ndjson` file
+    post:
+      operationId: ImportRules
+      x-codegen-enabled: false
+      summary: Import rules
+      description: Imports rules from an `.ndjson` file, including actions and exception lists.
+      tags:
+        - Import/Export API
+      requestBody:
+        required: true
+        content:
+          multipart/form-data:
+            schema:
+              type: object
+              properties:
+                file:
+                  type: string
+                  format: binary
+                  description: The `.ndjson` file containing the rules.
+      parameters:
+        - name: overwrite
+          in: query
+          required: false
+          description: Determines whether existing rules with the same `rule_id` are overwritten.
+          schema:
+            type: boolean
+            default: false
+        - name: overwrite_exceptions
+          in: query
+          required: false
+          description: Determines whether existing exception lists with the same `list_id` are overwritten.
+          schema:
+            type: boolean
+            default: false
+        - name: overwrite_action_connectors
+          in: query
+          required: false
+          description: Determines whether existing actions with the same `kibana.alert.rule.actions.id` are overwritten.
+          schema:
+            type: boolean
+            default: false
+      responses:
+        200:
+          description: Indicates a successful call.
+          content:
+            application/json:
+              schema:
+                type: object
