+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/model/query_rule_by_ids_validation.ts
@@ -10,11 +10,7 @@ import type { QueryRuleByIds } from './query_rule_by_ids';
 /**
  * Additional validation that is implemented outside of the schema itself.
  */
-export const validateQueryRuleByIds = (schema: QueryRuleByIds): string[] => {
-  return [...validateId(schema)];
-};
-
-const validateId = (rule: QueryRuleByIds): string[] => {
+export const validateQueryRuleByIds = (rule: QueryRuleByIds): string[] => {
   if (rule.id != null && rule.rule_id != null) {
     return ['both "id" and "rule_id" cannot exist, choose one or the other'];
   } else if (rule.id == null && rule.rule_id == null) {
