+++ x-pack/plugins/security_solution/common/api/detection_engine/rule_management/read_tags/read_tags_route.schema.yaml
@@ -1,0 +1,20 @@
+openapi: 3.0.0
+info:
+  title: Tags API endpoint
+  version: 2023-10-31
+paths:
+  /api/detection_engine/tags:
+    summary: Aggregates and returns rule tags
+    get:
+      operationId: GetTags
+      x-codegen-enabled: false
+      summary: Aggregates and returns all unique tags from all rules
+      tags:
+        - Tags API
+      responses:
+        200:
+          description: Indicates a successful call
+          content:
+            application/json:
+              schema:
+                $ref: '../../model/rule_schema/common_attributes.schema.yaml#/components/schemas/RuleTagArray'
