+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_management/api/rules/find_rules/route.ts
@@ -9,10 +9,7 @@ import type { IKibanaResponse, Logger } from '@kbn/core/server';
 import { transformError } from '@kbn/securitysolution-es-utils';
 
 import { DETECTION_ENGINE_RULES_URL_FIND } from '../../../../../../../common/constants';
-import type {
-  FindRulesRequestQueryDecoded,
-  FindRulesResponse,
-} from '../../../../../../../common/api/detection_engine/rule_management';
+import type { FindRulesResponse } from '../../../../../../../common/api/detection_engine/rule_management';
 import {
   FindRulesRequestQuery,
   validateFindRulesRequestQuery,
@@ -29,9 +26,7 @@ export const findRulesRoute = (router: SecuritySolutionPluginRouter, logger: Log
     {
       path: DETECTION_ENGINE_RULES_URL_FIND,
       validate: {
-        query: buildRouteValidation<typeof FindRulesRequestQuery, FindRulesRequestQueryDecoded>(
-          FindRulesRequestQuery
-        ),
+        query: buildRouteValidation(FindRulesRequestQuery),
       },
       options: {
         tags: ['access:securitySolution'],
